Final Project File
Performs image processing to determine which color is being detected. Maps the color to a measurement model to determine possible locations of the robot on the map. Uses Bayesian Localization to make the possible locations converge to a specific value. PD control is used for line following. All modules come together to make a mail delivery robot.